Move ComplexExpressionTypes to datatype package
